Move installation instructions to getting started
This makes the Getting Started guide the single (only) source of truth
for installing the metapackage components, moving the correct
information from the contributing guide to the getting started pages,
which were previously out-of-date.

This also splits the installation components of the getting started
pages into separate pages rather than separate tabs, because the
`..  tabbed::` directive didn't play nicely with Sphinx headings and
references, which caused all links to the source-installation sections
to be broken.